moved all to set.py rip set_blacklist.py and suggest.py
